Edited the model Schema for points into a double instead of integer and award 0.2 points for each eligible chat message

+++ db/migrations/add-column-points.js
@@ -1,9 +1,7 @@
-
-
 module.exports = {
   up: (queryInterface, Sequelize) =>
     queryInterface.addColumn('Members', 'points', {
-      type: Sequelize.INTEGER,
+      type: Sequelize.DOUBLE,
     }),
 
   down: (queryInterface, Sequelize) =>
