+++ db/models/member.js
@@ -7,7 +7,7 @@ module.exports = (sequelize, DataTypes) => {
     uuid: DataTypes.STRING,
     verified: DataTypes.BOOLEAN,
     messagesCount: DataTypes.INTEGER,
-    points: DataTypes.INTEGER,
+    points: DataTypes.DOUBLE,
 
   }, {
     classMethods: {
